smartos-live#2 Fixed identation, spaces should be tabs smartos-live#51 No symlink for libusb.so to libusb.so.1 for -lusb smartos-live#86 'make clean' fails if Node.js hasn't been built smartos-live#91 make world fails on trying to include zlib.h OS-1301 smartos-live subprojects should be told about the proto area OS-1302 smartos-live should offer a way to bootstrap illumos-extra OS-1303 smartos-live src_addon can die OS-1304 smartos-live src build should not use node-waf OS-1305 illumos-extra should not put libuuid.a into the proto area, ever OS-1306 non-bootstrappable runtime dependencies should be isolated OS-1310 smartos-live build doesn't fail even if kvm-cmd completely bombs out OS-1312 smartos-live makefiles must fit in 80 columns OS-1317 GZ crle configuration pollution

+++ src/bootparams.c
@@ -183,7 +183,7 @@ void raw_cat(int rfd)
         struct stat sbuf;
 
         if (fstat(wfd, &sbuf) == 0 &&
-            sbuf.st_blksize > sizeof(fb_buf)) {
+	    sbuf.st_blksize > (long)sizeof(fb_buf)) {
             bsize = sbuf.st_blksize;
             buf = malloc(bsize);
         }
